[Core] Make `makeEqnSolveHandler` return unique_ptr to the underlying handler
