+++ examples/EqnHandler/EqnHandler-MPI.cpp
@@ -54,7 +54,7 @@ int main(int argc, char* argv[]) {
             [&](auto&& e) { return d2x<D2SecondOrderCentered>(e) + d2y<D2SecondOrderCentered>(e) == 1.0; }, u,
             solver);
     OP_MPI_MASTER_INFO("Built solver handler.");
-    handler.solve();
+    handler->solve();
     OP_MPI_MASTER_INFO("Solver finished.");
 
     FinalizeEnvironment();
