+++ include/Core
@@ -100,9 +100,10 @@
 #include "Core/BC/ProxyBC.hpp"
 
 // Equation
-#include "Core/Equation/EqnSolveHandler.hpp"
+#include "Core/Equation/HYPREEqnSolveHandler.hpp"
 #include "Core/Equation/AMGCLEqnSolveHandler.hpp"
 #include "Core/Equation/Equation.hpp"
+#include "Core/Equation/EqnSolveHandler.hpp"
 #include "Core/Equation/UnifiedSolve.hpp"
 
 // Solvers
